+++ 05/src/main/java/ru/otus/spring05/service/IOService.java
@@ -4,5 +4,5 @@ public interface IOService {
 
     public void printOut(String st);
     public String readString() ;
-    public int readInt() ;
+    public Long readLong() ;
 }
